chore: move logger and _utils into utils submodule
