+++ rex_xai/box.py
@@ -10,7 +10,7 @@ from anytree import LevelOrderGroupIter, NodeMixin, RenderTree
 import numpy as np
 
 from rex_xai.distributions import Distribution, random_coords
-from rex_xai.logger import logger
+from rex_xai.utils.logger import logger
 
 # Enums for different axes
 class Axes:
