+++ rex_xai/database.py
@@ -11,7 +11,7 @@ from ast import literal_eval
 import pandas as pd
 import numpy as np
 
-from rex_xai.logger import logger
+from rex_xai.utils.logger import logger
 from rex_xai.config import CausalArgs, Strategy
 from rex_xai.explanation import Explanation
 from rex_xai.multi_explanation import MultiExplanation
