+++ rex_xai/onnx.py
@@ -15,7 +15,7 @@ from onnxruntime import InferenceSession
 from rex_xai.prediction import Prediction, from_pytorch_tensor
 from rex_xai.input_data import Setup
 
-from rex_xai.logger import logger
+from rex_xai.utils.logger import logger
 
 
 class OnnxRunner:
