+++ rex_xai/resp_maps.py
@@ -13,8 +13,8 @@ from rex_xai.box import Box
 from rex_xai.config import CausalArgs
 from rex_xai.mutant import Mutant
 from rex_xai.input_data import Data
-from rex_xai.logger import logger
-from rex_xai._utils import ReXMapError
+from rex_xai.utils.logger import logger
+from rex_xai.utils._utils import ReXMapError
 
 
 class ResponsibilityMaps:
