+++ rex_xai/rex_wrapper.py
@@ -1,11 +1,11 @@
 """main entry point to ReX"""
 
-from rex_xai._utils import get_device
+from rex_xai.utils._utils import get_device
 from rex_xai.config import get_all_args
 from rex_xai.database import initialise_rex_db
 
 from rex_xai.rex import explanation
-from rex_xai.logger import logger, set_log_level
+from rex_xai.utils.logger import logger, set_log_level
 from rex_xai.config import validate_args
