+++ rex_xai/visualisation.py
@@ -17,8 +17,8 @@ from rex_xai.prediction import Prediction
 from rex_xai.config import CausalArgs
 from rex_xai.resp_maps import ResponsibilityMaps
 from rex_xai.input_data import Data
-from rex_xai._utils import add_boundaries
-from rex_xai.logger import logger
+from rex_xai.utils._utils import add_boundaries
+from rex_xai.utils.logger import logger
 
 
 def plot_curve(curve, chunk_size, style="insertion", destination=None):
